kernel-port: xenia.exe LINKS - Edge kernel+vfs fully merged and building
~1100 kernel errors -> 0. xenia-kernel and xenia-app both build clean;
xenia.exe links (21.6MB, valid PE).

Root causes (each accounted for hundreds of errors):
- ~500 errors were ONE line: Edge's ui/imgui_drawer.h includes
  third_party/imgui/imgui.h where ours only forward-declared it, so every
  byte-identical xam/ui/*.cc dialog failed at once.
- xthread.h tested XE_PLATFORM_WIN32 BEFORE including platform.h, so
  <csetjmp> was silently skipped while the identical #if 620 lines later
  was true -> 'std::jmp_buf is not a member of std'.
- A std::filesystem::path passed to XELOGE (fmt 6.2 has no path
  formatter) made MSVC blame an unrelated log call 150 lines earlier.
- Three d3d12 errors were collateral from this port's C++20 switch
  (/std:c++20 implies /permissive-, which rejects &<rvalue>), not the
  kernel merge - fixed mechanically.

Third-party rule held: nothing vendored, nothing deleted. asio -> kept
master's raw-socket xsocket/xam_net; zlib-ng -> XLast gzip stubbed
(HasXLast() false, consumers already degrade); libavformat -> applied the
XE_XMP_HOST_DECODE_ENABLED gate that a previous step declared but never
used (this also fixed a real null-deref: Play() ignored the codec-open
result then dereferenced the context).

Latent bugs found and fixed: emulator.cc never populated
game_info_database_ (accessor would have null-dereffed on first title
name query); merge duplicates in xboxkrnl_crypt.cc (both trees'
XeKeysGetKey/XeCryptRotSumSha/XeCryptBnQwBeSigVerify), xsemaphore.h
(waiters_ twice) and xboxkrnl_io_info.cc (two XFileRenameInformation
cases) removed. gameinfo_utils.{h,cc} restored from master - the merge
dropped them while emulator.cc still used kernel::util::GameInfo.

Stubs, all NOTE(kernel-port) with re-enable paths: Emulator::
RelaunchTitle (in_process_title_relaunch defaults FALSE - the call sites
park the calling guest thread expecting in-process termination),
DataMigration, SaveGameConfigSetting (cpptoml vs tomlplusplus - live
cvars still update, only cross-launch persistence is lost),
DebugListener::OnDebugPrint, Processor::RemoveModule, hid portal, ui
title-font/SeparatorText.

+++ src/xenia/kernel/xam/xam_locale.cc
@@ -20,7 +20,10 @@
 #include "xenia/xbox.h"
 
 #include "third_party/fmt/include/fmt/format.h"
-#include "third_party/fmt/include/fmt/xchar.h"
+// NOTE(kernel-port): Edge also includes fmt/xchar.h (fmt 9+). Our vendored fmt
+// is 6.2.0, which has no xchar.h but already declares is_char<char16_t> in
+// core.h, so format.h alone covers the fmt::format(u"...") calls below.
+// Re-enable path: upgrade third_party/fmt, then restore the xchar.h include.
 
 // TODO(gibbed): put these forward decls in a header somewhere.
 
@@ -508,7 +511,8 @@ void XamFormatDateString_entry(dword_t locale_format, qword_t filetime,
   if (X_FILETIME(filetime).is_valid()) {
     XFormatDateString(filetime, output_buffer.guest_address(), output_count);
   } else {
-    XFormatDateString(ctx->r[3], ctx->r[4], ctx->r[5]);
+    XFormatDateString(ctx->r[3], static_cast<uint32_t>(ctx->r[4]),
+                      static_cast<uint32_t>(ctx->r[5]));
   }
 }
 DECLARE_XAM_EXPORT1(XamFormatDateString, kNone, kImplemented);
@@ -535,7 +539,8 @@ void XamFormatTimeString_entry(dword_t user_index, qword_t filetime,
   if (X_FILETIME(filetime).is_valid()) {
     XFormatTimeString(filetime, output_buffer.guest_address(), output_count);
   } else {
-    XFormatTimeString(ctx->r[3], ctx->r[4], ctx->r[5]);
+    XFormatTimeString(ctx->r[3], static_cast<uint32_t>(ctx->r[4]),
+                      static_cast<uint32_t>(ctx->r[5]));
   }
 }
 DECLARE_XAM_EXPORT1(XamFormatTimeString, kNone, kImplemented);
